IMPLEMENTACION DEL 30-03-25

+++ src/corlib/System/SByte.cs
@@ -4,7 +4,7 @@
     {
         public override string ToString()
         {
-            return ((long)this).ToString();
+            return ((int)this).ToString();
         }
     }
 }
